pyproject.toml
[project]
name = "ai-lawyer-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-ollama>=0.3.6",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
